+++ GodotTestDriver/Drivers/ItemListDriver.cs
@@ -96,7 +96,7 @@ namespace GodotTestDriver.Drivers
                 
                 uiControl.Select(i);
                 // calling this function will not emit the signal so we need to do this ourselves
-                uiControl.EmitSignal("item_selected", i);
+                uiControl.EmitSignal(ItemList.SignalName.ItemSelected, i);
                 await uiControl.GetTree().WaitForEvents();
                 return;
             }